AND-11 AND-12 WIP: Implemented DataChannel for chat on Web. DataChannel chat is sent in the format: CHAT|<Type of chat>|<Sender's nickname>|<chat data> For e.g.: CHAT|PRIVATE|webrtcUser|Chat data here, any character is allowed, including '|'... CHAT|GROUP|webrtcUser|Chat data here, any character is allowed, including '|'...
On Web client, to activate/inactivate Signalling channel or DataChannel chat,
set _chatSig or _chatDC to true or false respectively.
Both variables are set to true by default. A possible UI based controller implementation is to add a button or radio button for each variable to be toggled on and off.

Android client is still unable to receive chat in a reliable way from web client. This could be due to inability to activate SCTP channel on Android.